permissions for generating; hidden decomposition button and feedback

+++ client/scripts/tree.js
@@ -7,6 +7,8 @@ let last_update = null;
 let feedback_list = [];
 let expanded_tasks = [];
 
+let hide_implementation = true;
+
 const zoom = d3.zoom().on('zoom', function(e) {
     $('#task-data').modal('hide');
     g.attr('transform', e.transform);
@@ -214,7 +216,7 @@ function draw() {
         .classed('node-icon-implementation', true)
         .classed('fa', true)
         .classed('fa-code', true)
-        .classed('hidden', d => d.data.get_state() != 'implementable')
+        .classed('hidden', d => d.data.get_state() != 'implementable' && !hide_implementation)
         .on('click', open_node_menu)       // needed since it's on top of the circle
         .attr('width', 20)
         .attr('height', 20)
@@ -256,7 +258,7 @@ function draw() {
         .classed('hidden', d => !d.data.has_children())
         .on('click', (e, d) => d.data.is_leaf() ? show_children(d.data, d) : hide_children(d.data, d));
     nodesG_update.select('.node-icon-implementation')
-        .classed('hidden', d => d.data.get_state() != 'implementable')
+        .classed('hidden', d => d.data.get_state() != 'implementable' && !hide_implementation)
         .on('click', open_node_menu);       // needed since it's on top of the circle
 
     // Nodes - Exit
@@ -498,7 +500,7 @@ function show_buttons(item) {
 
     // Implement: only available on unsolved tasks that can be implemented
     let button_implement = $('#implement');
-    if (!item.data.is_solved() && item.data.can_be_implemented()) {
+    if (!item.data.is_solved() && item.data.can_be_implemented() && !hide_implementation) {
         button_implement.show();
         $('#dont-implement').unbind().on('click', function() {
             delete_implementation(item);
@@ -535,8 +537,15 @@ function generate_decomposition(item) {
 
     let task = item.data;
 
-    task.generate_decomposition(user_id, function() {
+    task.generate_decomposition(user_id, function(d) {
         task.running = false;
+
+        d = JSON.parse(d);
+
+        if (d.status == 'not_allowed') {
+            alert('This option is not enabled for your account');
+            return;
+        }
 
         // show this task after refresh
         expanded_tasks.push(task.task_id);
@@ -661,7 +670,16 @@ function generate_implementation(item, language, additional_instructions = null)
     if (!check_user_id())
         return;    
 
-    item.data.generate_implementation(user_id, language, additional_instructions, update, function(e) {
+    item.data.generate_implementation(user_id, language, additional_instructions, function(d) {
+        d = JSON.parse(d);
+
+        if (d.status == 'not_allowed') {
+            alert('This option is not enabled for your account');
+            return;
+        }
+
+        update();
+    }, function(e) {
         console.error(e);
         item.data.running = false;
         alert('error, see console for info');
